Restructure app to be more sensibly organised
