+++ tests/test_cs_transform.py
@@ -1,5 +1,5 @@
 from server import app
-from transform.views.test_views import test_message
+from app.views.test_views import test_message
 import unittest
 import io
 import zipfile
