--- transform/__init__.py
@@ -1,6 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-import transform.views.test_views  # noqa
-import transform.views.main  # noqa
